refactor: move all Taipy app code, data, and tests into taipy_implementation/ folder; update docs and requirements accordingly
