+++ README.md
@@ -1,2 +1,42 @@
 # ds-cs-toolkit
 A growing collection of mini-projects exploring data science and computer science tools. Each folder is a hands-on experiment, comparison, or deep dive into a specific topic, framework, or library.
+
+## Projects
+
+### taipy_implementation/
+A comprehensive sales performance dashboard built with Taipy, featuring interactive visualizations, data filtering, and real-time metrics calculation.
+
+- **Interactive Dashboard**: Real-time sales performance visualization
+- **Data Filtering**: Filter by date range and product categories  
+- **Multiple Visualizations**: Revenue over time, by category, and top products
+- **Comprehensive Testing**: 58 tests with 97% code coverage
+
+[View Taipy Implementation →](taipy_implementation/README.md)
+
+### dspy_implementation/
+DSPy-based implementation for natural language processing and AI applications.
+
+[View DSPy Implementation →](dspy_implementation/)
+
+## Getting Started
+
+Each project in this toolkit is self-contained with its own requirements and documentation. Navigate to the specific project folder to get started:
+
+```bash
+# For Taipy dashboard
+cd taipy_implementation
+pip install -r requirements.txt
+python taipy-app.py
+
+# For DSPy implementation  
+cd dspy_implementation
+# Follow project-specific instructions
+```
+
+## Contributing
+
+When adding new projects:
+1. Create a new folder for your implementation
+2. Include a comprehensive README.md
+3. Add appropriate tests and documentation
+4. Update this main README with project details
