+++ taipy_implementation/requirements.txt
@@ -1,0 +1,7 @@
+polars>=0.20.0
+numpy>=1.24.0 
+taipy>=4.1.0
+pytest>=7.4.0
+pytest-cov>=4.1.0
+pytest-mock>=3.11.0
+pandas>=2.0.0
